Fixed mysql connection
